+++ packages/perp/CHANGELOG.md
@@ -1,31 +1,67 @@
 # @orderly.network/futures
 
-## 3.3.9-dmm.2
+## 3.3.12
 
 ### Patch Changes
 
-- update
-- Updated dependencies
-  - @orderly.network/types@1.5.9-dmm.2
-  - @orderly.network/utils@1.5.9-dmm.2
+- 11399d1f3: add element id;
+- Updated dependencies [11399d1f3]
+  - @orderly.network/types@1.5.12
+  - @orderly.network/utils@1.5.12
 
-## 3.3.9-dmm.1
-
-### Patch Changes
-
-- update
-- Updated dependencies
-  - @orderly.network/types@1.5.9-dmm.1
-  - @orderly.network/utils@1.5.9-dmm.1
-
-## 3.3.9-dmm.0
+## 3.3.12-alpha.0
 
 ### Patch Changes
 
-- update
+- add element id;
 - Updated dependencies
-  - @orderly.network/types@1.5.9-dmm.0
-  - @orderly.network/utils@1.5.9-dmm.0
+  - @orderly.network/types@1.5.12-alpha.0
+  - @orderly.network/utils@1.5.12-alpha.0
+
+## 3.3.11
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/types@1.5.11
+  - @orderly.network/utils@1.5.11
+
+## 3.3.10
+
+### Patch Changes
+
+- e5a88d5f3: add system maintenance component;
+- Updated dependencies [e5a88d5f3]
+  - @orderly.network/types@1.5.10
+  - @orderly.network/utils@1.5.10
+
+## 3.3.10-alpha.0
+
+### Patch Changes
+
+- add system maintenance component;
+- Updated dependencies
+  - @orderly.network/types@1.5.10-alpha.0
+  - @orderly.network/utils@1.5.10-alpha.0
+
+## 3.3.9
+
+### Patch Changes
+
+- 187dd07b: fix bugs;
+- Updated dependencies [187dd07b]
+  - @orderly.network/types@1.5.9
+  - @orderly.network/utils@1.5.9
+
+## 3.3.9-alpha.0
+
+### Patch Changes
+
+- fix bugs;
+- Updated dependencies
+  - @orderly.network/types@1.5.9-alpha.0
+  - @orderly.network/utils@1.5.9-alpha.0
 
 ## 3.3.8